+++ R/launch_shiny.R
@@ -1,6 +1,0 @@
-# launch shiny app
-
-launch_shiny <- function(shiny_app) {
-  appDir <- system.file("shiny", shiny_app, package = "qlcData")
-  shiny::runApp(appDir, display.mode = "normal")
-}