+++ R/readwrite_recoding.R
@@ -1,79 +1,37 @@
-# ==========================================================
-# help function: expand values for combination of attributes
-# ==========================================================
-
-.expandValues <- function(attributes, data, all) {
-  
-  alldata <- data[, attributes, drop = FALSE]
-  alldata <- apply(alldata, 1, function(x){
-    paste(x, collapse = " + ")
-  })
-  freq <- table(alldata)
-  
-  if (all) {
-    combination <- expand.grid(
-      sapply( attributes, function(x){ 
-        c(levels(data[,x]),NA) 
-        }, simplify = FALSE )
-    )
-    combination <- apply(combination, 1, function(x){
-      paste(x, collapse = " + ")
-    })
-    freq <- freq[combination]
-    names(freq) <- combination
-    
-  }
-  return(as.list(freq))
-}
-
 # ===================
 # write YAML-template
 # ===================
 
-write.recoding <- function(data, attributes = NULL, all.options = FALSE, file = NULL, yaml = TRUE) {
+write.recoding <- function(data, attributes = NULL, all.options = FALSE, file = NULL) {
   
-  # prepare data when single column
+  # prepare data when vector
   if (is.null(dim(data))) {
     data <- as.data.frame(data)
     attributes <- 1
     colnames(data) <- 1
-  } else if (is.null(attributes)) {
+  }
+  # take all columns when not attributes are specified
+  if (is.null(attributes)) {
     attributes <- 1:ncol(data)
   }
-  
-  # prepare the template for one attribute
-  makeTemplate <- function(attribute, data) {
-    if (length(attribute) > 1) {
-      originalValues <- .expandValues(attribute, data, all = all.options)
-    } else {
-      originalValues <- as.list(table(data[,attribute]))
-    }
-    link <- sapply(originalValues, function(x){NULL})
-    return(list(
-      recodingOf = attribute,
-      attribute = NULL,
-      values = list(NULL, NULL),
-      link = link,
-      originalFrequency = originalValues,
-      comments = NULL
-    ))
-  }
+  # turn attributes into a list if necessary, replace numbers with names
+  attributes <- as.list(sapply(attributes, function(x) {
+    colnames(data[,x,drop = FALSE])
+  }))
+  names(attributes) <- NULL
   
   # combine all templates
-  attributes <- as.list(sapply(attributes, function(x){colnames(data)[x]}))
   result <- list(
     title = NULL,
     author = NULL,
     date = format(Sys.time(),"%Y-%m-%d"),
-    original_data = NULL,
-    recoding = sapply(attributes, function(x) { makeTemplate(x, data) }, simplify = FALSE)
+    originalData = NULL,
+    selectRows = NULL,
+    recoding = sapply(attributes, function(x) { .makeTemplate(x, data, all.options) }, simplify = FALSE)
   )
   
-  # return the result, defaults to a yaml-file
-  if (yaml) {
-    if (is.null(file)) {
-      stop("please specify file")
-    }
+  # return the result
+  if (!is.null(file)) {
     yaml <- yaml::as.yaml(result)
     yaml <- gsub("\n- recodingOf:","\n# ==========\n- recodingOf:",yaml)
     cat(yaml, file = file)
@@ -103,103 +61,26 @@ read.recoding <- function(recoding, file = NULL, data = NULL) {
     }
   }
   
-  # Allow for various shortcuts in the writing of recodings
-  # The following lines normalise the input to the cannonical form
-  
-  reallabels <- c(
-    "recodingOf", "attribute", "values", "link",
-    "originalFrequency", "doNotRecode", "comments")
-  remove <- c()
-  
-  # This loop has become a mess, should be cleaned up!
+  # This is the central loop expanding all recodings
   for (i in 1:length(recoding)) {
-    
-    # write labels in full
-    names(recoding[[i]]) <- reallabels[pmatch(names(recoding[[i]]),reallabels)]    
-    
-    # when doNotRecode is specified, you're ready to go
-    if (is.null(recoding[[i]]$doNotRecode)) {
-      # if not: then recodingOf is necessary, otherwise break
-      if (is.null(recoding[[i]]$recodingOf)) {
-        stop(paste("Specify **recodingOf** for recoding number", i))
-      }
-      # with no link, add doNotRecode
-      if (length(unlist(recoding[[i]]$link)) == 0) { 
-        recoding[[i]] <- list(doNotRecode = recoding[[i]]$recodingOf)
-      } else {
- 
-        # prepare values
-        if (is.null(recoding[[i]]$values)) {
-          # make values from link
-          newValues <- levels(factor(unlist(recoding[[i]]$link)))
-          recoding[[i]]$values <- newValues
-          names(recoding[[i]]$values) <- newValues
-        } else {
-          recoding[[i]]$values <- unlist(recoding[[i]]$values)
-        }
-        
-        # prepare link
-        if (is.null(names(recoding[[i]]$values))) {
-          # when values have no labels, interpret link as numeric
-          recoding[[i]]$link <- unlist(recoding[[i]]$link)
-        } else {
-          # when values have labels, interpret link as character
-          recoding[[i]]$link <- sapply(recoding[[i]]$link, as.character)
-        }
-        
-        # just for visual inspection
-        recoding[[i]]$originalFrequency <- unlist(recoding[[i]]$originalFrequency)
-        
-        # make attribute and value names if necessary
-        if (is.null(recoding[[i]]$attribute)) {
-          recoding[[i]]$attribute <- paste0("Att", i)
-        }
-        if (is.null(unlist(recoding[[i]]$values))) {
-          recoding[[i]]$values <- paste0("val", 1:length(recoding[[i]]$link))
-        }
-      }
-    } else {
-      # break on possible error
-      if (!is.null(recoding[[i]]$link)) {
-        stop(paste("Both doNotRecode and link specified in recoding number", i))
-      }
-    }
-    
-    # when data is specified, add names of original attributes
-    # this leads to nicer documentation of the recoding
-    if (!is.null(data)) {
-      if (is.numeric(recoding[[i]]$recodingOf)) {
-        recoding[[i]]$recodingOf <- colnames(data)[recoding[[i]]$recodingOf]
-      }
-      if (is.numeric(recoding[[i]]$doNotRecode)) {
-        recoding[[i]]$doNotRecode <- colnames(data)[recoding[[i]]$doNotRecode]
-      }    
-    }
-    # put everything in the same order
-    recoding[[i]] <- recoding[[i]][reallabels]
-    recoding[[i]] <- recoding[[i]][na.omit(names(recoding[[i]]))]
-    
-    # merge sequences of doNotRecode
-    if (i > 1) {
-      if (!is.null(recoding[[i]]$doNotRecode) & 
-          !is.null(recoding[[i-1]]$doNotRecode)) {
-        recoding[[i]]$doNotRecode <- c( recoding[[i-1]]$doNotRecode
-                                       , recoding[[i]]$doNotRecode)
-        remove <- c(remove,(i-1))
-      }
-    }
-  }
-  
-  # remove superflous recodings because of contractions of doNotRecode
-  if (!is.null(remove)) {
-    recoding <- recoding[-remove]
+    recoding[[i]] <- .doExpandRecoding(recoding[[i]], counter = i, data = data)
   }
   
   # return result
   if (is.null(file)) {
+    # check for data-address in profile
+    if (!is.null(meta$originalData)) {
+      recoding$originalData <- meta$originalData
+    }
     return(recoding)
   } else {
     # add metadata and write out as yaml
+    if (!("selectRows" %in% names(meta))) {
+      meta <- c(list(selectRows = NULL), meta)
+    }    
+    if (!("originalData" %in% names(meta))) {
+      meta <- c(list(originalData = deparse(substitute(data))), meta)
+    }
     if (!("date" %in% names(meta))) {
       meta <- c(list(date = format(Sys.time(),"%Y-%m-%d")), meta)
     }
@@ -215,4 +96,164 @@ read.recoding <- function(recoding, file = NULL, data = NULL) {
     yaml <- gsub("\n- doNotRecode:","\n# ==========\n- doNotRecode:",yaml)
     cat(yaml, file = file)
   }
+}
+
+# ==========================================================
+# help function: prepare the template for one attribute
+# ==========================================================
+
+.makeTemplate <- function(attribute, data, all.options) {
+  if (length(attribute) > 1) {
+    originalValues <- .expandValues(attribute, data, all = all.options)
+  } else {
+    originalValues <- as.list(table(data[,attribute]))
+  }
+  link <- sapply(originalValues, function(x){NULL})
+  return(list(
+    recodingOf = attribute,
+    attribute = NULL,
+    values = list(a = NULL,b = NULL),
+    link = link,
+    newData = NULL,
+    originalFrequency = originalValues,
+    comments = NULL
+  ))
+}
+
+# ==========================================================
+# help function: expand values for combination of attributes
+# ==========================================================
+
+.expandValues <- function(attributes, data, all) {
+  
+  alldata <- data[, attributes, drop = FALSE]
+  alldata <- apply(alldata, 1, function(x){
+    paste(x, collapse = " + ") # This separator is currently hard coded!
+  })
+  freq <- table(alldata)
+  
+  if (all) {
+    combination <- expand.grid(
+      sapply( attributes, function(x){ 
+        c(levels(data[,x]),NA) 
+      }, simplify = FALSE )
+    )
+    combination <- apply(combination, 1, function(x){
+      paste(x, collapse = " + ")
+    })
+    freq <- freq[combination]
+    names(freq) <- combination
+    
+  }
+  return(as.list(freq))
+}
+
+# ======================================================
+# help function: expand shortcuts and normalize recoding
+# =======================================================
+
+.doExpandRecoding <- function(recoding, counter, data) {
+  
+  # =============
+  # preliminaries
+  # =============  
+  
+  # Allow for various shortcuts in the writing of recodings
+  # The following lines normalise the input to the cannonical form
+  reallabels <- c(
+    "recodingOf", "attribute", "values", "link", "newData",
+    "originalFrequency", "doNotRecode", "comments")
+  
+  # write labels in full
+  names(recoding) <- reallabels[pmatch(names(recoding),reallabels)]    
+  
+  # make template if data is available
+  if (!is.null(data)) {
+    template <- write.recoding(data = data, attributes = recoding$recodingOf)
+    template <- template$recoding[[1]]
+  }
+  
+  # when doNotRecode is specified, possibly add names
+  if (!is.null(recoding$doNotRecode)) {
+    if (!is.null(data)) {
+      recoding$doNotRecode <- colnames(data[,recoding$doNotRecode, drop = FALSE])
+    }
+    # but break on possible error
+    if (!is.null(recoding$link)) {
+      stop(paste("Both **doNotRecode** and **link** specified in recoding number"
+                 , counter))
+    }
+  } else {
+    # if no doNotRecode: then recodingOf is necessary, otherwise break
+    if (is.null(recoding$recodingOf)) {
+      stop(paste("Specify **recodingOf** for recoding number", counter))
+    }
+    
+    # =============================================     
+    # here starts the real expanding of information
+    # =============================================
+    
+    # with no link, add doNotRecode
+    if (length(unlist(recoding$link)) == 0) { 
+      recoding <- list(doNotRecode = recoding$recodingOf)
+      if (!is.null(data)) {
+        recoding$doNotRecode <- template$recodingOf
+      }
+    } else {
+      
+      # if recodingOf is number and data is available, look up names
+      if (is.numeric(recoding$recodingOf) & !is.null(data)) {
+        recoding$recodingOf <- template$recodingOf
+      }
+      
+      # make new attribute name if left blanc
+      if (is.null(recoding$attribute)) {
+        recoding$attribute <- paste0("Att", counter)
+      }    
+      
+      # prepare link
+      recoding$link <- unlist(recoding$link)
+      # numeric link: turn into integer
+      if (is.numeric(recoding$link)) {
+        recoding$link <- as.integer(recoding$link)
+      }
+      # when no names for link given, assume levels of data
+      if (!is.null(data) & is.null(names(recoding$link))) {
+        names(recoding$link) <- names(template$link)
+      }
+      
+      # prepare values if not present
+      if (is.null(unlist(recoding$values))) {
+        # make new values when none specified and link is numeric
+        if (is.numeric(recoding$link)) {
+          recoding$values <- paste0("val", 1:length(unique(recoding$link)))
+        } else {
+          # make values from link
+          recoding$values <- levels(factor(recoding$link))
+        } 
+      }
+      
+      # add frequencies from data if specified
+      # or copy data from profile
+      # will be left empty if neither is available
+      if (!is.null(data)) {
+        recoding$originalFrequency <- template$originalFrequency
+      } else {
+        recoding$originalFrequency <- unlist(recoding$originalFrequency)
+      }      
+    }
+  }
+  
+  # =======
+  # cleanup
+  # =======  
+  
+  # for yaml output: turn into lists
+  recoding$link <- as.list(recoding$link)
+  recoding$values <- as.list(recoding$values)
+  # put everything in the same order
+  recoding <- recoding[reallabels]
+  # remove superfluous
+  recoding <- recoding[sapply(recoding, length) > 0]
+  
 }