+++ NEWS.md
@@ -1,5 +1,13 @@
+# qlcData 0.3
+
+* recoding now tries to read the data from originalData if not specified
+* read.profile now includes lines starting with a hash '#'
+* adding options for selectRows and addData to recoding
+* various minor edits to recoding functions
+
 # qlcData 0.2
 
+* update instructions to link docopt executables
 * changing default recoding profile to empirically attested combinations of features
 * add glottolog data with access function getTree() and asPhylo()
 * add possibility for factors in recoding